Allow securing specific resources in GraphQL (#1595)
